@@ -8,7 +8,8 @@
     "sync-models": "npx epi-sync-models -e development",
     "test": "echo \"Error: no test specified\" && exit 1",
     "clean": "echo \"Clean: not implemented\" && exit 0",
-    "build": "npx webpack --env EPI_ENV=development --progress",
+    "start": "npx webpack serve --env EPI_ENV=development --config webpack.dev.config.js",
+    "build": "npx webpack --env EPI_ENV=development",
     "build:production": "npx epi-sync-models -e production && npx webpack --env EPI_ENV=production && npx epi-sync-models -e development",
     "watch": "webpack --watch --progress --env EPI_ENV=development",
     "compile": "tsc --pretty --project ./tsconfig.json"
@@ -16,7 +17,8 @@
   "author": "Remko Jantzen",
   "license": "Apache-2.0",
   "dependencies": {
-    "@episerver/spa-core": "git+https://github.com/episerver/foundation-lib-spa-core.git#Develop",
+    "@episerver/foundation-settings": "file:lib/foundation-settings",
+    "@episerver/spa-core": "git+https://github.com/episerver/foundation-lib-spa-core.git#0.1.0",
     "@fullcalendar/daygrid": "^5.4.0",
     "@fullcalendar/list": "^5.4.0",
     "@fullcalendar/react": "^5.4.0",
@@ -27,10 +29,12 @@
     "eventemitter3": "^4.0.7",
     "html-entities": "^2.0.6",
     "jquery": "^3.5.1",
+    "leaflet": "^1.7.1",
     "lodash": "^4.17.19",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-helmet": "^6.0.0",
+    "react-leaflet": "^3.1.0",
     "react-player": "^2.8.1",
     "react-redux": "^7.2.2",
     "react-router": "^5.2.0",
@@ -39,11 +43,12 @@
     "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@episerver/webpack": "git+https://github.com/episerver/foundation-lib-spa-webpack.git#Develop",
+    "@episerver/webpack": "git+https://github.com/episerver/foundation-lib-spa-webpack.git#0.1.0",
     "@fortawesome/fontawesome-free": "^5.15.2",
     "@types/bootstrap": "^4.6.0",
     "@types/html-entities": "^1.2.16",
     "@types/jquery": "^3.5.5",
+    "@types/leaflet": "^1.5.23",
     "@types/lodash": "^4.14.168",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
@@ -53,29 +58,31 @@
     "@types/react-router-dom": "^5.1.7",
     "@types/uuid": "^8.3.0",
     "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^7.0.0",
+    "copy-webpack-plugin": "^8.0.0",
     "css-loader": "^5.0.1",
     "dotenv": "^8.2.0",
     "dotenv-expand": "^5.1.0",
     "esm": "^3.2.25",
     "file-loader": "^6.2.0",
-    "filemanager-webpack-plugin": "^3.1.0",
-    "html-webpack-plugin": "^4.5.1",
+    "filemanager-webpack-plugin": "^4.0.0",
+    "html-webpack-plugin": "^5.1.0",
     "loader-utils": "^2.0.0",
     "mini-css-extract-plugin": "^1.3.4",
     "node-sass": "^5.0.0",
-    "postcss-loader": "^4.2.0",
+    "postcss-loader": "^5.0.0",
     "precss": "^4.0.0",
     "resolve-url-loader": "^3.1.2",
-    "sass-loader": "^10.1.1",
+    "sass-loader": "^11.0.1",
     "schema-utils": "^3.0.0",
     "source-map-loader": "^2.0.0",
     "style-loader": "^2.0.0",
     "terser-webpack-plugin": "^5.1.1",
     "ts-loader": "^8.0.14",
-    "typescript": "^4.1.3",
+    "typescript": "^4.2.2",
     "webpack": "^5.17.0",
     "webpack-cli": "^4.4.0",
+    "webpack-dev-server": "^3.11.2",
+    "webpack-manifest-plugin": "^3.0.0",
     "webpack-merge": "^5.7.3",
     "webpack-node-externals": "^2.5.2",
     "xmlhttprequest": "^1.8.0",
