@@ -1,15 +1,14 @@
 ﻿using EPiServer.Core;
 using EPiServer.Framework.DataAnnotations;
 using EPiServer.Web.Mvc;
-using Foundation.Cms.Blocks;
 using System.Web.Mvc;
 
 namespace Foundation.ContentDelivery.Controller
 {
     [TemplateDescriptor(Inherited = true)]
-    public class DefaultBlockController : BlockController<FoundationBlockData>
+    public class DefaultBlockController : BlockController<BlockData>
     {
-        public override ActionResult Index(FoundationBlockData currentBlock)
+        public override ActionResult Index(BlockData currentBlock)
         {
             return PartialView(currentBlock);
         }
