@@ -5,8 +5,6 @@ using EPiServer.Framework.Web;
 using EPiServer.Framework.Web.Mvc;
 using EPiServer.Web;
 using EPiServer.Web.Mvc;
-using Foundation.Cms.Pages;
-using System.Linq;
 using System.Web.Mvc;
 
 namespace Foundation.ContentDelivery.Controller
@@ -20,66 +18,9 @@ namespace Foundation.ContentDelivery.Controller
     [RequireClientResources]
     public class DefaultPreviewController : ActionControllerBase, IRenderTemplate<BlockData>
     {
-        private readonly IContentLoader _contentLoader;
-        private readonly TemplateResolver _templateResolver;
-        private readonly DisplayOptions _displayOptions;
-
-        public DefaultPreviewController(IContentLoader contentLoader, TemplateResolver templateResolver, DisplayOptions displayOptions)
-        {
-            _contentLoader = contentLoader;
-            _templateResolver = templateResolver;
-            _displayOptions = displayOptions;
-        }
-
         public ActionResult Index(IContent currentContent)
         {
-            //As the layout requires a page for title etc we "borrow" the start page
-            var startPage = _contentLoader.Get<CmsHomePage>(ContentReference.StartPage);
-
-            var model = new DefaultPreviewModel(startPage, currentContent);
-
-            var supportedDisplayOptions = _displayOptions
-                .Select(x => new
-                {
-                    x.Tag,
-                    x.Name,
-                    Supported = SupportsTag(currentContent, x.Tag)
-                }).ToList();
-
-            if (!supportedDisplayOptions.Any(x => x.Supported))
-            {
-                return View(model);
-            }
-            foreach (var displayOption in supportedDisplayOptions)
-            {
-                var contentArea = new ContentArea();
-                contentArea.Items.Add(new ContentAreaItem
-                {
-                    ContentLink = currentContent.ContentLink
-                });
-                var areaModel = new DefaultPreviewModel.PreviewArea
-                {
-                    Supported = displayOption.Supported,
-                    AreaTag = displayOption.Tag,
-                    AreaName = displayOption.Name,
-                    ContentArea = contentArea
-                };
-
-                model.Areas.Add(areaModel);
-            }
-
-            return View(model);
-        }
-
-        private bool SupportsTag(IContent content, string tag)
-        {
-            var templateModel = _templateResolver.Resolve(HttpContext,
-                content.GetOriginalType(),
-                content,
-                TemplateTypeCategories.MvcPartial,
-                tag);
-
-            return templateModel != null;
+            return View(currentContent);
         }
     }
 }
