@@ -6,11 +6,15 @@ using EPiServer.ContentApi.Core;
 using EPiServer.ContentApi.Core.Configuration;
 using EPiServer.ContentApi.Routing;
 using EPiServer.ContentApi.Search;
+using EPiServer.Core;
 using EPiServer.ServiceLocation;
-using Foundation.Cms.Extensions;
+using EPiServer.Web;
+using EPiServer.Web.Routing;
 using Foundation.ContentDelivery.Infrastructure;
 using Foundation.ContentDelivery.Models;
 using Foundation.ContentDelivery.Models.TypeConverters;
+using Microsoft.Owin.Cors;
+using Owin;
 using System;
 using System.Collections.Generic;
 using System.Configuration;
@@ -23,12 +27,24 @@ namespace Foundation.ContentDelivery
 {
     public static class Extensions
     {
+
+        private static readonly Lazy<IUrlResolver> UrlResolver =
+            new Lazy<IUrlResolver>(() => ServiceLocator.Current.GetInstance<IUrlResolver>());
+
+        private static readonly Lazy<ISiteDefinitionResolver> SiteDefinitionResolver =
+            new Lazy<ISiteDefinitionResolver>(() => ServiceLocator.Current.GetInstance<ISiteDefinitionResolver>());
+
+        public static IAppBuilder UseContentApiCors(this IAppBuilder app)
+        {
+            app.UseCors(new CorsOptions() { });
+            return app;
+        }
+
         public static void ConfigureFoundationContentDelivery(this ServiceConfigurationContext context)
         {
-            context.ConfigureFoundationCms();
 
             var corsString = ConfigurationManager.AppSettings["foundation:contentDelivery:cors_Urls"];
-            if (corsString.IsNullOrEmpty() && File.Exists(HttpContext.Current.Server.MapPath("/") + "..\\Spa.FrontEnd\\.env"))
+            if (string.IsNullOrEmpty(corsString) && File.Exists(HttpContext.Current.Server.MapPath("/") + "..\\Spa.FrontEnd\\.env"))
             {
                 context.Services.AddSingleton<IEnvReader, EnvReader>();
                 AddEnv(builder =>
@@ -51,7 +67,7 @@ namespace Foundation.ContentDelivery
             context.Services.Configure<ContentApiConfiguration>(c =>
             {
                 var clients = new List<ContentApiClient>();
-                if (corsString.IsNullOrEmpty())
+                if (string.IsNullOrEmpty(corsString))
                 {
 #if DEBUG
                     clients.Add(new ContentApiClient { ClientId = "Default", AccessControlAllowOrigin = "*" });
@@ -96,6 +112,34 @@ namespace Foundation.ContentDelivery
 
             var dotEnvOptions = dotEnvOptionsBuilder.Build();
             DotEnv.Config(dotEnvOptions);
+        }
+
+        /// <summary>
+        /// Helper method to get a URL string for a content reference using the provided culture code
+        /// </summary>
+        /// <param name="contentRef">The content reference of a routable content item to get the URL for.</param>
+        /// <param name="lang">The language code to use when retrieving the URL.</param>
+        /// <param name="isAbsolute">Whether the full URL including protocol and host should be returned.</param>
+        public static Uri GetUri(this ContentReference contentRef, string lang, bool isAbsolute = false)
+        {
+            var urlString = UrlResolver.Value.GetUrl(contentRef, lang, new UrlResolverArguments { ForceCanonical = true });
+            if (string.IsNullOrEmpty(urlString))
+            {
+                return new Uri(string.Empty);
+            }
+
+            //if we're not getting an absolute URL, we don't need to work out the correct host name so exit here
+            var uri = new Uri(urlString, UriKind.RelativeOrAbsolute);
+            if (uri.IsAbsoluteUri || !isAbsolute)
+            {
+                return uri;
+            }
+
+            //Work out the correct domain to use from the hosts defined in the site definition
+            var siteDefinition = SiteDefinitionResolver.Value.GetByContent(contentRef, true, true);
+            var host = siteDefinition.Hosts.FirstOrDefault(h => h.Type == HostDefinitionType.Primary) ?? siteDefinition.Hosts.FirstOrDefault(h => h.Type == HostDefinitionType.Undefined);
+            var baseUrl = (host?.Name ?? "*").Equals("*") ? siteDefinition.SiteUrl : new Uri($"http{((host.UseSecureConnection ?? false) ? "s" : string.Empty)}://{host.Name}");
+            return new Uri(baseUrl, urlString);
         }
     }
 }
