@@ -1,27 +1,29 @@
 ﻿@using EPiServer.Framework.Web.Resources
 @using EPiServer.Shell
 @using EPiServer.Shell.Web.Mvc.Html
+
 <!DOCTYPE html>
 <html lang="en">
 <head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta content='width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no' name='viewport'>
-    <link rel="icon" href="~/Assets/icons/episerver.png" type="image/x-icon" />
-    <link rel="shortcut icon" href="~/Assets/icons/episerver.png" type="image/x-icon" />
+    <link rel="icon" href="@Paths.ToClientResource("Foundation.Cms", "ClientResources/img/episerver.png")" type="image/x-icon" />
+    <link rel="shortcut icon" href="@Paths.ToClientResource("Foundation.Cms", "ClientResources/img/episerver.png")" type="image/x-icon" />
     @Html.Raw(ClientResources.RenderResources("ShellCore"))
     @Html.Raw(ClientResources.RenderResources("ShellWidgets"))
     @Html.Raw(ClientResources.RenderResources("ShellCoreLightTheme"))
     @Html.Raw(ClientResources.RenderResources("Navigation"))
     @Html.Raw(ClientResources.RenderResources("DojoDashboardCompatibility", new[] { ClientResourceType.Style }))
-    <link href="~/Assets/scss/vendors/font-awesome-5.9.0/css/all.css" rel="stylesheet" type="text/css" />
-    <link href="@Paths.ToClientResource("Foundation.Cms", "ClientResources/css/components/grid-icons.css")" rel="stylesheet" type="text/css" />
-    <link href="~/Assets/scss/vendors/spur-admin/spur.min.css" rel="stylesheet" type="text/css" />
-    <link href="~/Assets/scss/main.min.css" rel="stylesheet" type="text/css" />
-    <link href="~/Assets/scss/vendors/mvc-grid/mvc-grid.css" rel="stylesheet" type="text/css" />
+    <link href="@Paths.ToClientResource("Foundation.Cms", "ClientResources/css/vendors/bootstrap.min.css")" rel="stylesheet" type="text/css" />
+    <link href="@Paths.ToClientResource("Foundation.Cms", "ClientResources/css/vendors/font-awesome-5.9.0/css/all.min.css")" rel="stylesheet" type="text/css" />
+    <link href="@Paths.ToClientResource("Foundation.Cms", "ClientResources/css/features/grid-icons.css")" rel="stylesheet" type="text/css" />
+    <link href="@Paths.ToClientResource("Foundation.Cms", "ClientResources/css/vendors/spur.min.css")" rel="stylesheet" type="text/css" />
+    @*<link href="~/Assets/scss/main.min.css" rel="stylesheet" type="text/css" />*@
+    <link href="@Paths.ToClientResource("Foundation.Cms", "ClientResources/css/vendors/mvc-grid/mvc-grid.css")" rel="stylesheet" type="text/css" />
     @RenderSection("AdditionalStyles", false)
     <style>
-        body > .main-container, body > .container, body > .dash-content  {
+        body > .main-container, body > .container, body > .dash-content {
             padding-top: 50px;
         }
 
@@ -39,18 +41,16 @@
         </div>
         @RenderBody()
     </div>
-    <script src="~/Assets/js/vendors/jquery/jquery-3.4.0.min.js" type="text/javascript"></script>
-    <script src="~/Assets/js/vendors/popper/popper.min.js" type="text/javascript"></script>
-    <script src="~/Assets/js/vendors/bootstrap-4.3.1/bootstrap.min.js" type="text/javascript"></script>
-    <script src="~/Assets/js/vendors/spur-admin/spur.js" type="text/javascript"></script>
-    <script src="~/Assets/js/vendors/mvc-grid/mvc-grid.js" type="text/javascript"></script>
-    <script src="~/Assets/js/vendors/feather-icons/feather.min.js" type="text/javascript"></script>
-    <script src="~/Assets/js/vendors/axios/axios.min.js" type="text/javascript"></script>
-    <script src="~/Assets/js/plugins/notify.js" type="text/javascript"></script>
-    <script src="~/Assets/js/features/notifyHelper.js" type="text/javascript"></script>
+    <script src="@Paths.ToClientResource("Foundation.Cms", "ClientResources/js/vendors/jquery-3.4.0.min.js")" type="text/javascript"></script>
+    <script src="@Paths.ToClientResource("Foundation.Cms", "ClientResources/js/vendors/popper.min.js")" type="text/javascript"></script>
+    <script src="@Paths.ToClientResource("Foundation.Cms", "ClientResources/js/vendors/bootstrap.min.js")" type="text/javascript"></script>
+    <script src="@Paths.ToClientResource("Foundation.Cms", "ClientResources/js/vendors/spur.js")" type="text/javascript"></script>
+    <script src="@Paths.ToClientResource("Foundation.Cms", "ClientResources/js/vendors/mvc-grid.js")" type="text/javascript"></script>
+    <script src="@Paths.ToClientResource("Foundation.Cms", "ClientResources/js/vendors/axios.min.js")" type="text/javascript"></script>
+    <script src="@Paths.ToClientResource("Foundation.Cms", "ClientResources/js/vendors/notify.min.js")" type="text/javascript"></script>
+    <script src="@Paths.ToClientResource("Foundation.Cms", "ClientResources/js/features/notify-helper.js")" type="text/javascript"></script>
     <script type="text/javascript">
         $(document).ready(function () {
-            feather.replace()
             axios.defaults.headers.common['Accept'] = '*/*';
             window.notification = new NotifyHelper();
         });
