@@ -3,4 +3,4 @@ module.exports = {
         require('precss'),
         require('autoprefixer')
     ]
-  }
+}
