@@ -11,10 +11,17 @@ export default class TextboxElementBlock extends AbstractElementBlock<TextboxEle
         let fieldName : string = fieldId;
         let fieldInfoId : string = `${ fieldId }_Description`;
 
+        let label = null;
+        let description = null;
+        if (this.props.data?.label?.value) 
+            label = <label htmlFor={ fieldId }><Components.Property iContent={this.props.data} field="label" context={ this.getContext() }/></label>;
+        if (this.props.data?.description?.value)
+            description = <small id={ fieldInfoId } className="form-text text-muted"><Components.Property iContent={this.props.data} field="description" context={ this.getContext() }/></small>;
+
         return <div className="form-group">
-            <label htmlFor={ fieldId }><Components.Property iContent={this.props.data} field="label" context={ this.getContext() }/></label>
-            <input type={ this.getFieldType() } id={ fieldId } name={ fieldName } className="form-control" defaultValue={ this.props.data.predefinedValue?.value } placeholder={ this.props.data.placeHolder?.value } aria-describedby={ fieldInfoId } />
-            <small id={ fieldInfoId } className="form-text text-muted"><Components.Property iContent={this.props.data} field="description" context={ this.getContext() }/></small>
+            { label }
+            <input type={ this.getFieldType() } id={ fieldId } name={ fieldName } className="form-control" defaultValue={ this.props.data.predefinedValue?.value } placeholder={ this.props.data.placeHolder?.value } aria-describedby={ description ? fieldInfoId : undefined } />
+            { description }            
         </div>
     }
 
