@@ -3,73 +3,77 @@ import { Taxonomy, Core, Loaders } from '@episerver/spa-core';
 export default class TypeMapper extends Loaders.BaseTypeMapper {
   protected map : { [type: string]: Loaders.TypeInfo } = {
     'CookieDropBlock': {dataModel: 'CookieDropBlockData',instanceModel: 'CookieDropBlockType'},
-    'BlogItemPage': {dataModel: 'BlogItemPageData',instanceModel: 'BlogItemPageType'},
-    'LocationListPage': {dataModel: 'LocationListPageData',instanceModel: 'LocationListPageType'},
-    'LocationItemPage': {dataModel: 'LocationItemPageData',instanceModel: 'LocationItemPageType'},
+    'ResetPasswordMailPage': {dataModel: 'ResetPasswordMailPageData',instanceModel: 'ResetPasswordMailPageType'},
+    'PersonPage': {dataModel: 'PersonPageData',instanceModel: 'PersonPageType'},
+    'PersonList': {dataModel: 'PersonListData',instanceModel: 'PersonListType'},
+    'ProfileLandingPage': {dataModel: 'ProfileLandingPageData',instanceModel: 'ProfileLandingPageType'},
+    'SearchResultPage': {dataModel: 'SearchResultPageData',instanceModel: 'SearchResultPageType'},
+    'SearchSettings': {dataModel: 'SearchSettingsData',instanceModel: 'SearchSettingsType'},
+    'ReferencePageSettings': {dataModel: 'ReferencePageSettingsData',instanceModel: 'ReferencePageSettingsType'},
+    'LayoutSettings': {dataModel: 'LayoutSettingsData',instanceModel: 'LayoutSettingsType'},
+    'ResetPasswordPage': {dataModel: 'ResetPasswordPageData',instanceModel: 'ResetPasswordPageType'},
+    'LabelSettings': {dataModel: 'LabelSettingsData',instanceModel: 'LabelSettingsType'},
+    'StandardPage': {dataModel: 'StandardPageData',instanceModel: 'StandardPageType'},
+    'SpaMedia': {dataModel: 'SpaMediaData',instanceModel: 'SpaMediaType'},
     'SpaFolder': {dataModel: 'SpaFolderData',instanceModel: 'SpaFolderType'},
     'CategoryRoot': {dataModel: 'CategoryRootData',instanceModel: 'CategoryRootType'},
     'CustomViewConfigurationBlock': {dataModel: 'CustomViewConfigurationBlockData',instanceModel: 'CustomViewConfigurationBlockType'},
-    'ContentRecommendationsBlock': {dataModel: 'ContentRecommendationsBlockData',instanceModel: 'ContentRecommendationsBlockType'},
-    'YouTubeBlock': {dataModel: 'YouTubeBlockData',instanceModel: 'YouTubeBlockType'},
-    'VimeoBlock': {dataModel: 'VimeoBlockData',instanceModel: 'VimeoBlockType'},
-    'VideoBlock': {dataModel: 'VideoBlockData',instanceModel: 'VideoBlockType'},
-    'TwitterBlock': {dataModel: 'TwitterBlockData',instanceModel: 'TwitterBlockType'},
-    'TextBlock': {dataModel: 'TextBlockData',instanceModel: 'TextBlockType'},
-    'TeaserBlock': {dataModel: 'TeaserBlockData',instanceModel: 'TeaserBlockType'},
-    'RssReaderBlock': {dataModel: 'RssReaderBlockData',instanceModel: 'RssReaderBlockType'},
-    'RecentPageCategoryRecommendationBlock': {dataModel: 'RecentPageCategoryRecommendationBlockData',instanceModel: 'RecentPageCategoryRecommendationBlockType'},
-    'PageListBlock': {dataModel: 'PageListBlockData',instanceModel: 'PageListBlockType'},
-    'NavigationBlock': {dataModel: 'NavigationBlockData',instanceModel: 'NavigationBlockType'},
-    'BlogListPage': {dataModel: 'BlogListPageData',instanceModel: 'BlogListPageType'},
-    'MenuItemBlock': {dataModel: 'MenuItemBlockData',instanceModel: 'MenuItemBlockType'},
-    'BookmarksPage': {dataModel: 'BookmarksPageData',instanceModel: 'BookmarksPageType'},
-    'CmsHomePage': {dataModel: 'CmsHomePageData',instanceModel: 'CmsHomePageType'},
-    'SpaMedia': {dataModel: 'SpaMediaData',instanceModel: 'SpaMediaType'},
+    'BreadcrumbBlock': {dataModel: 'BreadcrumbBlockData',instanceModel: 'BreadcrumbBlockType'},
+    'CallToActionBlock': {dataModel: 'CallToActionBlockData',instanceModel: 'CallToActionBlockType'},
+    'ButtonBlock': {dataModel: 'ButtonBlockData',instanceModel: 'ButtonBlockType'},
+    'CollectionSettings': {dataModel: 'CollectionSettingsData',instanceModel: 'CollectionSettingsType'},
+    'CarouselBlock': {dataModel: 'CarouselBlockData',instanceModel: 'CarouselBlockType'},
+    'ProfilePage': {dataModel: 'ProfilePageData',instanceModel: 'ProfilePageType'},
+    'CodingFile': {dataModel: 'CodingFileData',instanceModel: 'CodingFileType'},
     'PdfFile': {dataModel: 'PdfFileData',instanceModel: 'PdfFileType'},
+    'SettingsFolder': {dataModel: 'SettingsFolderData',instanceModel: 'SettingsFolderType'},
+    'BlogItemPage': {dataModel: 'BlogItemPageData',instanceModel: 'BlogItemPageType'},
+    'BlogListPage': {dataModel: 'BlogListPageData',instanceModel: 'BlogListPageType'},
     'StandardCategory': {dataModel: 'StandardCategoryData',instanceModel: 'StandardCategoryType'},
+    'CalendarEventPage': {dataModel: 'CalendarEventPageData',instanceModel: 'CalendarEventPageType'},
+    'FolderPage': {dataModel: 'FolderPageData',instanceModel: 'FolderPageType'},
+    'HomePage': {dataModel: 'HomePageData',instanceModel: 'HomePageType'},
+    'BookmarksPage': {dataModel: 'BookmarksPageData',instanceModel: 'BookmarksPageType'},
+    'LandingPage': {dataModel: 'LandingPageData',instanceModel: 'LandingPageType'},
+    'TwoColumnLandingPage': {dataModel: 'TwoColumnLandingPageData',instanceModel: 'TwoColumnLandingPageType'},
+    'LocationItemPage': {dataModel: 'LocationItemPageData',instanceModel: 'LocationItemPageType'},
+    'LocationListPage': {dataModel: 'LocationListPageData',instanceModel: 'LocationListPageType'},
+    'TagPage': {dataModel: 'TagPageData',instanceModel: 'TagPageType'},
     'VideoFile': {dataModel: 'VideoFileData',instanceModel: 'VideoFileType'},
     'StandardFile': {dataModel: 'StandardFileData',instanceModel: 'StandardFileType'},
     'ImageMediaData': {dataModel: 'ImageMediaDataData',instanceModel: 'ImageMediaDataType'},
     'FoundationPdfFile': {dataModel: 'FoundationPdfFileData',instanceModel: 'FoundationPdfFileType'},
-    'TwoColumnLandingPage': {dataModel: 'TwoColumnLandingPageData',instanceModel: 'TwoColumnLandingPageType'},
     'ThreeColumnLandingPage': {dataModel: 'ThreeColumnLandingPageData',instanceModel: 'ThreeColumnLandingPageType'},
-    'TagPage': {dataModel: 'TagPageData',instanceModel: 'TagPageType'},
-    'StandardPage': {dataModel: 'StandardPageData',instanceModel: 'StandardPageType'},
-    'SearchResultPage': {dataModel: 'SearchResultPageData',instanceModel: 'SearchResultPageType'},
-    'ResetPasswordPage': {dataModel: 'ResetPasswordPageData',instanceModel: 'ResetPasswordPageType'},
-    'ResetPasswordMailPage': {dataModel: 'ResetPasswordMailPageData',instanceModel: 'ResetPasswordMailPageType'},
-    'ProfilePage': {dataModel: 'ProfilePageData',instanceModel: 'ProfilePageType'},
-    'LandingPage': {dataModel: 'LandingPageData',instanceModel: 'LandingPageType'},
-    'FolderPage': {dataModel: 'FolderPageData',instanceModel: 'FolderPageType'},
-    'CalendarEventPage': {dataModel: 'CalendarEventPageData',instanceModel: 'CalendarEventPageType'},
+    'GoogleMapsBlock': {dataModel: 'GoogleMapsBlockData',instanceModel: 'GoogleMapsBlockType'},
+    'ContainerBlock': {dataModel: 'ContainerBlockData',instanceModel: 'ContainerBlockType'},
     'HeroBlock': {dataModel: 'HeroBlockData',instanceModel: 'HeroBlockType'},
-    'LinkedInCompanyBlock': {dataModel: 'LinkedInCompanyBlockData',instanceModel: 'LinkedInCompanyBlockType'},
-    'GoogleMapsEmbedBlock': {dataModel: 'GoogleMapsEmbedBlockData',instanceModel: 'GoogleMapsEmbedBlockType'},
-    'HeroBlockCallout': {dataModel: 'HeroBlockCalloutData',instanceModel: 'HeroBlockCalloutType'},
+    'HealthChatbotBlock': {dataModel: 'HealthChatbotBlockData',instanceModel: 'HealthChatbotBlockType'},
     'CustomizedSearchAdvancedSettings': {dataModel: 'CustomizedSearchAdvancedSettingsData',instanceModel: 'CustomizedSearchAdvancedSettingsType'},
     'CustomizedSearchSettings': {dataModel: 'CustomizedSearchSettingsData',instanceModel: 'CustomizedSearchSettingsType'},
     'CustomizedSearchBlock': {dataModel: 'CustomizedSearchBlockData',instanceModel: 'CustomizedSearchBlockType'},
     'FilterActivitiesBlock': {dataModel: 'FilterActivitiesBlockData',instanceModel: 'FilterActivitiesBlockType'},
-    'FilterContinentsBlock': {dataModel: 'FilterContinentsBlockData',instanceModel: 'FilterContinentsBlockType'},
     'FilterDistancesBlock': {dataModel: 'FilterDistancesBlockData',instanceModel: 'FilterDistancesBlockType'},
-    'ExistsFilterBlock': {dataModel: 'ExistsFilterBlockData',instanceModel: 'ExistsFilterBlockType'},
-    'NumericFilterBlock': {dataModel: 'NumericFilterBlockData',instanceModel: 'NumericFilterBlockType'},
     'FilterTemperaturesBlock': {dataModel: 'FilterTemperaturesBlockData',instanceModel: 'FilterTemperaturesBlockType'},
-    'BreadcrumbBlock': {dataModel: 'BreadcrumbBlockData',instanceModel: 'BreadcrumbBlockType'},
-    'ButtonBlock': {dataModel: 'ButtonBlockData',instanceModel: 'ButtonBlockType'},
     'CalendarBlock': {dataModel: 'CalendarBlockData',instanceModel: 'CalendarBlockType'},
-    'CallToActionBlock': {dataModel: 'CallToActionBlockData',instanceModel: 'CallToActionBlockType'},
-    'CarouselBlock': {dataModel: 'CarouselBlockData',instanceModel: 'CarouselBlockType'},
-    'ContainerBlock': {dataModel: 'ContainerBlockData',instanceModel: 'ContainerBlockType'},
+    'YouTubeBlock': {dataModel: 'YouTubeBlockData',instanceModel: 'YouTubeBlockType'},
+    'VimeoBlock': {dataModel: 'VimeoBlockData',instanceModel: 'VimeoBlockType'},
+    'FilterContinentsBlock': {dataModel: 'FilterContinentsBlockData',instanceModel: 'FilterContinentsBlockType'},
+    'TwitterBlock': {dataModel: 'TwitterBlockData',instanceModel: 'TwitterBlockType'},
+    'TextBlock': {dataModel: 'TextBlockData',instanceModel: 'TextBlockType'},
+    'TeaserBlock': {dataModel: 'TeaserBlockData',instanceModel: 'TeaserBlockType'},
+    'RssReaderBlock': {dataModel: 'RssReaderBlockData',instanceModel: 'RssReaderBlockType'},
+    'PageListBlock': {dataModel: 'PageListBlockData',instanceModel: 'PageListBlockType'},
     'FacebookBlock': {dataModel: 'FacebookBlockData',instanceModel: 'FacebookBlockType'},
-    'StringFilterBlock': {dataModel: 'StringFilterBlockData',instanceModel: 'StringFilterBlockType'},
-    'BlogCommentBlock': {dataModel: 'BlogCommentBlockData',instanceModel: 'BlogCommentBlockType'},
+    'NavigationBlock': {dataModel: 'NavigationBlockData',instanceModel: 'NavigationBlockType'},
+    'MenuItemBlock': {dataModel: 'MenuItemBlockData',instanceModel: 'MenuItemBlockType'},
+    'HeroBlockCallout': {dataModel: 'HeroBlockCalloutData',instanceModel: 'HeroBlockCalloutType'},
+    'VideoBlock': {dataModel: 'VideoBlockData',instanceModel: 'VideoBlockType'},
     'TextboxElementBlock': {dataModel: 'TextboxElementBlockData',instanceModel: 'TextboxElementBlockType'},
     'TextareaElementBlock': {dataModel: 'TextareaElementBlockData',instanceModel: 'TextareaElementBlockType'},
     'NumberElementBlock': {dataModel: 'NumberElementBlockData',instanceModel: 'NumberElementBlockType'},
     'DateTimeRangeElementBlock': {dataModel: 'DateTimeRangeElementBlockData',instanceModel: 'DateTimeRangeElementBlockType'},
-    'DateTimeElementBlock': {dataModel: 'DateTimeElementBlockData',instanceModel: 'DateTimeElementBlockType'},
     'AddressesElementBlock': {dataModel: 'AddressesElementBlockData',instanceModel: 'AddressesElementBlockType'},
+    'DateTimeElementBlock': {dataModel: 'DateTimeElementBlockData',instanceModel: 'DateTimeElementBlockType'},
     'RangeElementBlock': {dataModel: 'RangeElementBlockData',instanceModel: 'RangeElementBlockType'},
     'UrlElementBlock': {dataModel: 'UrlElementBlockData',instanceModel: 'UrlElementBlockType'},
     'SelectionElementBlock': {dataModel: 'SelectionElementBlockData',instanceModel: 'SelectionElementBlockType'},
