@@ -1,8 +1,10 @@
 //Episerver library
-import Start from '@episerver/spa-core';
+import Start, { Core } from '@episerver/spa-core';
 
 //Application
 import config from 'app/Config';
 import 'app/styles.scss';
 
 Start(config);
+
+Core.DefaultContext.events().debug = false;
