@@ -9,10 +9,13 @@ const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const TerserPlugin = require('terser-webpack-plugin');
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 const FileManagerPlugin = require('filemanager-webpack-plugin');
+const { WebpackManifestPlugin } = require('webpack-manifest-plugin');
 
 // Epi Webpack tools
 const EpiWebpack = require('@episerver/webpack');
 
+// Configuration
+const manifest = require('./manifest.json');
 
 module.exports = (env) => {
     // Bundle info
@@ -23,6 +26,7 @@ module.exports = (env) => {
     const epiEnv        = env.EPI_ENV || process.env.EPI_ENV;
 	const config        = new EpiWebpack.Config(__dirname, env, epiEnv);
     const srcPath       = config.getSourceDir();
+    const pubPath       = config.getAssetPath();
 
     // Environment configs
     const webPath       = config.getWebPath();
@@ -30,6 +34,8 @@ module.exports = (env) => {
     const forProduction = mode.toLowerCase() === 'production';
     const epiBaseUrl    = config.getEpiserverURL();
     const epiDeployPath = config.getEnvVariable('EPI_DEPLOY_PATH', '/api/episerver/v3/deploy');
+    const outPath       = webPath + 'spaview/' + bundle + '/';
+    const distPath      = path.join(__dirname, 'dist', bundle);
 
     const webpackConfig = {
         entry: path.resolve(srcPath,'index.tsx'),
@@ -39,12 +45,10 @@ module.exports = (env) => {
 		output: {
 			filename: 'scripts/[name].[contenthash:8].js',
 			chunkFilename: 'scripts/[name].[contenthash:8].js',
-			path: path.join(__dirname, 'dist', bundle),
-			publicPath: webPath + 'spaview/' + bundle + '/'
+			path: distPath,
+			publicPath: outPath
         },
-        resolve: config.getResolveConfig({
-            extensions: ['.css']
-        }),
+        resolve: config.getResolveConfig(),
         module: {
             rules: [
                 {
@@ -69,16 +73,57 @@ module.exports = (env) => {
                     loader: "source-map-loader"
                 },
                 {
-                    test: /\.(woff(2)?|ttf|eot|svg)(\?v=\d+\.\d+\.\d+)?$/,
+                    test: /\.(woff(2)?|ttf|eot)(\?v=\d+\.\d+\.\d+)?$/,
                     use: [
                         {
                             loader: 'file-loader',
                             options: {
                                 name: '[name].[ext]',
-                                outputPath: 'fonts'
+                                outputPath: 'fonts',
+                                publicPath: outPath + 'fonts/'
                             }
                         }
                     ]
+                },
+                {
+                    test: /\.(png|svg|jpg|jpeg)(\?v=\d+\.\d+\.\d+)?$/,
+                    use: [
+                        {
+                            loader: 'file-loader',
+                            options: {
+                                name: '[name].[ext]',
+                                outputPath: 'images/',
+                                publicPath: outPath + 'images/'
+                            }
+                        }
+                    ]
+                },
+                {
+                    test: /\.css$/,
+                    use: 
+                    [
+                        {
+                            loader: MiniCssExtractPlugin.loader,
+                            options: {
+                                publicPath: 'styles'
+                            }
+                        },
+                        {
+                            loader: 'css-loader',
+                            options: { 
+                                sourceMap: true 
+                            }
+                        }, 
+                        'postcss-loader',
+                        {
+                            loader: 'resolve-url-loader',
+                            options: {
+                                debug: true,
+                                root: outPath,
+                                sourceMap: true
+                            }
+                        }
+                    ],
                 },
                 {
                     test: /\.(s[ca]ss)$/,
@@ -93,19 +138,9 @@ module.exports = (env) => {
                             options: { 
                                 sourceMap: true 
                             }
-                        }, {
-                            loader: 'postcss-loader', // Run post css actions
-                            options: {
-                                postcssOptions: {
-                                    plugins: function () { // post css plugins, can be exported to postcss.config.js
-                                        return [
-                                            require('precss'),
-                                            require('autoprefixer')
-                                        ];
-                                    }
-                                }
-                            }
-                        }, {
+                        }, 
+                        'postcss-loader',
+                        {
                             loader: 'resolve-url-loader',
                             options: { 
                                 sourceMap: true
@@ -117,18 +152,9 @@ module.exports = (env) => {
                             }
                         }
                     ]
-                },
-                {
-                    test: /\.css$/,
-                    use: [ {
-                        loader: 'style-loader'
-                    }, {
-                        loader: 'css-loader'
-                    } ]
                 }
             ]
         },
-
         // Optimize frontend bundling
 		optimization: {
 			mergeDuplicateChunks: true,
@@ -136,7 +162,7 @@ module.exports = (env) => {
             emitOnErrors: false,
 			splitChunks: {
 				chunks: 'all',
-				maxInitialRequests: 1000,
+				maxInitialRequests: 50,
 				maxAsyncRequests: 1000,
 				minSize: 1,
 				automaticNameDelimiter: '.',
@@ -151,7 +177,8 @@ module.exports = (env) => {
 
 							// npm package names are URL-safe, but some servers don't like @ symbols
 							return `${cacheGroupKey}.${packageName.replace('@', '')}`;
-						}
+						},
+                        
                     },
                     
                     // Split Application components into separate files, might be needed if you don't provide a loader
@@ -161,6 +188,21 @@ module.exports = (env) => {
 							// get the name. E.g. node_modules/packageName/not/this/part.js
                             // or node_modules/packageName
                             const componentId = module.identifier().match(/[\\/]src[\\/][Cc]omponents[\\/](.*)/)[1].split(path.sep).map(x => x.split(".")[0]).join('.');
+
+							//const packageName = module.context.match(/[\\/][Cc]omponents[\\/](.*?)([\\/]|$)/)[1];
+
+							// npm package names are URL-safe, but some servers don't like @ symbols
+							return `${cacheGroupKey}.${componentId.toLowerCase().replace('@', '')}`;
+                        }
+					},
+                    
+                    // Split Application components into separate files, might be needed if you don't provide a loader
+					async_components: {
+						test: /[\\/]src[\\/][Aa]sync[Cc]omponents[\\/]/,
+						name(module, chunks, cacheGroupKey) {
+							// get the name. E.g. node_modules/packageName/not/this/part.js
+                            // or node_modules/packageName
+                            const componentId = module.identifier().match(/[\\/]src[\\/][Aa]sync[Cc]omponents[\\/](.*)/)[1].split(path.sep).map(x => x.split(".")[0]).join('.');
 
 							//const packageName = module.context.match(/[\\/][Cc]omponents[\\/](.*?)([\\/]|$)/)[1];
 
@@ -180,14 +222,21 @@ module.exports = (env) => {
 
             new HtmlWebpackPlugin({
                 template: 'src/index.html',
-                title: 'Episerver Foundation Single Page Application',
+                title: manifest.name,
                 filename: 'index.html',
+                packagePath: outPath,
                 minify: {
                     removeComments: false,
                     preserveLineBreaks: true,
                     collapseWhitespace: false,
                     collapseBooleanAttributes: true
                 }
+            }),
+
+            new WebpackManifestPlugin({
+                basePath: outPath,
+                writeToFileEmit: true,
+                seed: manifest
             }),
 
             new MiniCssExtractPlugin({
@@ -200,14 +249,8 @@ module.exports = (env) => {
                 {
                     patterns: [
                         {
-                            from: path.join(srcPath,'favicon.ico'),
-                            to: 'favicon.ico'
-                        }, {
-                            from: path.join(srcPath,'robots.txt'),
-                            to: 'robots.txt'
-                        }, {
-                            from: path.join(srcPath,'web.config'),
-                            to: 'web.config'
+                            from: pubPath,
+                            to: distPath.replace(path.sep, '/')+'/'
                         }
                     ]
                 }
@@ -242,6 +285,18 @@ module.exports = (env) => {
             })
         ]
     }
+
+    // Inject support for Leaflet package style loading
+    /*webpackConfig.resolve.extensions.push('.scss','.css');
+    webpackConfig.resolve.alias = { 
+        ...webpackConfig.resolve.alias, 
+        leaflet_css: path.join(__dirname, "/node_modules/leaflet/dist/leaflet.css"),
+        leaflet_marker: path.join(__dirname, "/node_modules/leaflet/dist/images/marker-icon.png"),
+        leaflet_marker_2x: path.join(__dirname, "/node_modules/leaflet/dist/images/marker-icon-2x.png"),
+        leaflet_marker_shadow: path.join(__dirname, "/node_modules/leaflet/dist/images/marker-shadow.png")
+    };*/
+    //console.log(webpackConfig.resolve);
+    //process.exit(0);
 
     return webpackConfig;
 }
